@@ -21,61 +21,73 @@
   },
   "private": true,
   "dependencies": {
-    "@agm/core": "^1.0.0-beta.0",
-    "@angular/cli": "1.0.0-rc.2",
-    "@angular/compiler-cli": "^2.4.0",
-    "@angular/common": "^2.4.0",
-    "@angular/compiler": "^2.4.0",
-    "@angular/core": "^2.4.0",
-    "@angular/forms": "^2.4.0",
-    "@angular/http": "^2.4.0",
-    "@angular/platform-browser": "^2.4.0",
-    "@angular/platform-browser-dynamic": "^2.4.0",
-    "@angular/router": "^3.4.0",
-    "animate.css": "^3.5.2",
-    "arrive": "^2.3.1",
-    "bootstrap": "^3.3.5",
-    "bootstrap-notify": "^3.1.3",
-    "bootstrap-select": "^1.12.2",
-    "bootstrap-tagsinput": "^0.7.1",
-    "chartist": "^0.9.4",
-    "chartist-plugin-zoom": "^0.4.0",
-    "core-js": "^2.4.1",
+    "@agm/core": "1.0.0-beta.0",
+    "@angular/animations": "4.3.1",
+    "@angular/cdk": "^2.0.0-beta.8",
+    "@angular/cli": "^1.2.5",
+    "@angular/common": "4.3.1",
+    "@angular/compiler": "4.3.1",
+    "@angular/compiler-cli": "4.3.1",
+    "@angular/core": "4.3.1",
+    "@angular/forms": "4.3.1",
+    "@angular/http": "4.3.1",
+    "@angular/material": "^2.0.0-beta.8",
+    "@angular/platform-browser": "4.3.1",
+    "@angular/platform-browser-dynamic": "4.3.1",
+    "@angular/platform-server": "4.3.1",
+    "@angular/router": "4.3.1",
+    "angular2-material-datepicker": "^0.5.0",
+    "animate.css": "3.5.2",
+    "arrive": "2.3.1",
+    "bootstrap": "3.3.5",
+    "bootstrap-material-design": "^0.5.10",
+    "bootstrap-notify": "3.1.3",
+    "bootstrap-select": "1.12.2",
+    "bootstrap-tagsinput": "0.7.1",
+    "chartist": "0.9.4",
+    "chartist-plugin-zoom": "0.4.0",
+    "core-js": "2.4.1",
     "datatables": "1.10.12",
     "datatables.net-bs": "1.10.12",
-    "datatables.net-responsive": "^2.1.1",
-    "domready": "^1.0.8",
+    "datatables.net-responsive": "2.1.1",
     "eonasdan-bootstrap-datetimepicker": "4.17.47",
     "express": "^4.15.3",
-    "fullcalendar": "^3.4.0",
-    "googleapis": "^19.0.0",
+    "fullcalendar": "3.4.0",
+    "googleapis": "19.0.0",
     "jasny-bootstrap": "^3.1.3",
-    "jquery": "^1.12.4",
-    "nouislider": "^9.2.0",
-    "rxjs": "^5.1.0",
-    "twitter-bootstrap-wizard": "^1.2.0",
-    "validate": "^3.0.1",
-    "web-animations-js": "^2.2.2",
-    "zone.js": "^0.7.6",
-    "@types/bootstrap": "^3.3.32",
-    "@types/chartist": "^0.9.34",
-    "@types/jasmine": "2.5.38",
-    "@types/jquery": "^1.10.31",
-    "@types/node": "^6.0.73"
+    "jquery": "1.12.4",
+    "moment": "^2.18.1",
+    "ng2-nouislider": "1.6.1",
+    "ng2-select": "^1.2.0",
+    "ngx-chips": "^1.4.6",
+    "nouislider": "9.2.0",
+    "rxjs": "5.4.2",
+    "twitter-bootstrap-wizard": "1.2.0",
+    "typescript": "2.4.2",
+    "validate": "3.0.1",
+    "web-animations-js": "2.2.2",
+    "zone.js": "0.8.4"
   },
   "devDependencies": {
-    "codelyzer": "~2.0.0",
-    "jasmine-core": "~2.5.2",
-    "jasmine-spec-reporter": "~3.2.0",
-    "karma": "~1.4.1",
-    "karma-chrome-launcher": "~2.0.0",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^0.2.0",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.0",
-    "ts-node": "~2.0.0",
-    "tslint": "~4.5.0",
-    "typescript": "~2.0.0"
+    "@angular/cli": "^1.2.5",
+    "@angular/compiler-cli": "2.4.0",
+    "@types/bootstrap": "3.3.32",
+    "@types/chartist": "0.9.34",
+    "@types/jasmine": "2.5.38",
+    "@types/jquery": "1.10.31",
+    "@types/node": "6.0.73",
+    "codelyzer": "2.0.0",
+    "jasmine-core": "2.5.2",
+    "jasmine-spec-reporter": "3.2.0",
+    "karma": "1.4.1",
+    "karma-chrome-launcher": "2.0.0",
+    "karma-cli": "1.0.1",
+    "karma-coverage-istanbul-reporter": "0.2.0",
+    "karma-jasmine": "1.1.0",
+    "karma-jasmine-html-reporter": "0.2.2",
+    "protractor": "5.1.0",
+    "ts-node": "2.0.0",
+    "tslint": "4.5.0",
+    "typescript": "2.4.2"
   }
 }
