@@ -1,129 +1,127 @@
-import { Component, OnInit, OnChanges, AfterViewInit } from '@angular/core';
-// import initWizard = require('../../../../assets/js/init/initWizard.js');
+import { Component, OnInit, OnChanges, AfterViewInit, SimpleChanges } from '@angular/core';
 
-declare var $:any;
+declare const $: any;
 interface FileReaderEventTarget extends EventTarget {
-    result:string
+    result: string;
 }
 interface FileReaderEvent extends Event {
     target: FileReaderEventTarget;
-    getMessage():string;
+    getMessage(): string;
 }
 @Component({
-    moduleId: module.id,
-    selector: 'wizard-cmp',
+    selector: 'app-wizard-cmp',
     templateUrl: 'wizard.component.html'
 })
 
-export class WizardComponent implements OnInit, OnChanges, AfterViewInit{
+export class WizardComponent implements OnInit, OnChanges, AfterViewInit {
     readURL(input) {
         if (input.files && input.files[0]) {
-            var reader = new FileReader();
+            const reader = new FileReader();
 
-            reader.onload = function (e:FileReaderEvent) {
+            reader.onload = function (e: FileReaderEvent) {
                 $('#wizardPicturePreview').attr('src', e.target.result).fadeIn('slow');
-            }
+            };
             reader.readAsDataURL(input.files[0]);
         }
     }
-    ngOnInit(){
+    ngOnInit() {
         // Code for the Validator
-        var $validator = $('.wizard-card form').validate({
-    		  rules: {
-    		    firstname: {
-    		      required: true,
-    		      minlength: 3
-    		    },
-    		    lastname: {
-    		      required: true,
-    		      minlength: 3
-    		    },
-    		    email: {
-    		      required: true,
-    		      minlength: 3,
-    		    }
+        const $validator = $('.wizard-card form').validate({
+            rules: {
+                  firstname: {
+                    required: true,
+                  minlength: 3
+              },
+                lastname: {
+                    required: true,
+                  minlength: 3
+              },
+                email: {
+                    required: true,
+                  minlength: 3,
+              }
             },
 
             errorPlacement: function(error, element) {
                 $(element).parent('div').addClass('has-error');
              }
-    	});
+         });
 
         // Wizard Initialization
-      	$('.wizard-card').bootstrapWizard({
+        $('.wizard-card').bootstrapWizard({
             'tabClass': 'nav nav-pills',
             'nextSelector': '.btn-next',
             'previousSelector': '.btn-previous',
 
             onNext: function(tab, navigation, index) {
-            	var $valid = $('.wizard-card form').valid();
-            	if(!$valid) {
-            		$validator.focusInvalid();
-            		return false;
-            	}
+                const $valid = $('.wizard-card form').valid();
+                if (!$valid) {
+                    $validator.focusInvalid();
+                    return false;
+                }
             },
 
-            onInit : function(tab, navigation, index){
+            onInit: function(tab, navigation, index){
 
-              //check number of tabs and fill the entire row
-              var $total = navigation.find('li').length;
-              var  $width = 100/$total;
-              var $wizard = navigation.closest('.wizard-card');
+              // check number of tabs and fill the entire row
+              const $total = navigation.find('li').length;
+              let  $width = 100 / $total;
+              const $wizard = navigation.closest('.wizard-card');
 
-              var $display_width = $(document).width();
+              const $display_width = $(document).width();
 
-              if($display_width < 600 && $total > 3){
+              if ($display_width < 600 && $total > 3) {
                   $width = 50;
               }
 
-               navigation.find('li').css('width',$width + '%');
-               var $first_li = navigation.find('li:first-child a').html();
-               var $moving_div = $('<div class="moving-tab">' + $first_li + '</div>');
+               navigation.find('li').css('width', $width + '%');
+               const $first_li = navigation.find('li:first-child a').html();
+               const $moving_div = $('<div class="moving-tab">' + $first_li + '</div>');
                $('.wizard-card .wizard-navigation').append($moving_div);
 
-            //    this.refreshAnimation($wizard, index);
-            var total_steps = $wizard.find('li').length;
-            var move_distance = $wizard.width() / total_steps;
-            var step_width = move_distance;
+            // this.refreshAnimation($wizard, index);
+            const total_steps = $wizard.find('li').length;
+            let move_distance = $wizard.width() / total_steps;
+            const step_width = move_distance;
             move_distance *= index;
 
-            var $current = index + 1;
+            const $current = index + 1;
 
-            if($current == 1){
+            if ($current === 1) {
                 move_distance -= 8;
-            } else if($current == total_steps){
+            } else if ($current === total_steps) {
                 move_distance += 8;
             }
 
             $wizard.find('.moving-tab').css('width', step_width);
             $('.moving-tab').css({
-                'transform':'translate3d(' + move_distance + 'px, 0, 0)',
+                'transform': 'translate3d(' + move_distance + 'px, 0, 0)',
                 'transition': 'all 0.5s cubic-bezier(0.29, 1.42, 0.79, 1)'
 
             });
 
-               $('.moving-tab').css('transition','transform 0s');
+               $('.moving-tab').css('transition', 'transform 0s');
            },
 
             onTabClick : function(tab, navigation, index){
 
-                var $valid = $('.wizard-card form').valid();
+                const $valid = $('.wizard-card form').valid();
 
-                if(!$valid){
+                if (!$valid) {
                     return false;
-                } else{
+                } else {
                     return true;
                 }
             },
 
             onTabShow: function(tab, navigation, index) {
-                var $total = navigation.find('li').length;
-                var $current = index+1;
+                const $total = navigation.find('li').length;
+                let $current = index + 1;
 
-                var $wizard = navigation.closest('.wizard-card');
+                const $wizard = navigation.closest('.wizard-card');
 
                 // If it's the last tab then hide the last button and show the finish instead
-                if($current >= $total) {
+                if ($current >= $total) {
                     $($wizard).find('.btn-next').hide();
                     $($wizard).find('.btn-finish').show();
                 } else {
@@ -131,74 +129,70 @@ export class WizardComponent implements OnInit, OnChanges, AfterViewInit{
                     $($wizard).find('.btn-finish').hide();
                 }
 
-                var button_text = navigation.find('li:nth-child(' + $current + ') a').html();
+                const button_text = navigation.find('li:nth-child(' + $current + ') a').html();
 
                 setTimeout(function(){
                     $('.moving-tab').text(button_text);
                 }, 150);
 
-                var checkbox = $('.footer-checkbox');
+                const checkbox = $('.footer-checkbox');
 
-                if( index !== 0 ){
+                if ( index !== 0 ) {
                     $(checkbox).css({
-                        'opacity':'0',
-                        'visibility':'hidden',
-                        'position':'absolute'
+                        'opacity': '0',
+                        'visibility': 'hidden',
+                        'position': 'absolute'
                     });
                 } else {
                     $(checkbox).css({
-                        'opacity':'1',
-                        'visibility':'visible'
+                        'opacity': '1',
+                        'visibility': 'visible'
                     });
                 }
 
                 // this.refreshAnimation($wizard, index);
-                var total_steps = $wizard.find('li').length;
-                var move_distance = $wizard.width() / total_steps;
-                var step_width = move_distance;
+                const total_steps = $wizard.find('li').length;
+                let move_distance = $wizard.width() / total_steps;
+                const step_width = move_distance;
                 move_distance *= index;
 
-                var $current = index + 1;
+                $current = index + 1;
 
-                if($current == 1){
+                if ($current === 1) {
                     move_distance -= 8;
-                } else if($current == total_steps){
+                } else if ($current === total_steps) {
                     move_distance += 8;
                 }
 
                 $wizard.find('.moving-tab').css('width', step_width);
                 $('.moving-tab').css({
-                    'transform':'translate3d(' + move_distance + 'px, 0, 0)',
+                    'transform': 'translate3d(' + move_distance + 'px, 0, 0)',
                     'transition': 'all 0.5s cubic-bezier(0.29, 1.42, 0.79, 1)'
-
                 });
             }
-      	});
+        });
 
 
         // Prepare the preview for profile picture
-        $("#wizard-picture").change(function(){
-
+        $('#wizard-picture').change(function(){
             this.readURL(this);
         });
 
         $('[data-toggle="wizard-radio"]').click(function(){
-            console.log('click');
-
-            var wizard = $(this).closest('.wizard-card');
+            const wizard = $(this).closest('.wizard-card');
             wizard.find('[data-toggle="wizard-radio"]').removeClass('active');
             $(this).addClass('active');
             $(wizard).find('[type="radio"]').removeAttr('checked');
-            $(this).find('[type="radio"]').attr('checked','true');
+            $(this).find('[type="radio"]').attr('checked', 'true');
         });
 
         $('[data-toggle="wizard-checkbox"]').click(function(){
-            if( $(this).hasClass('active')){
+            if ( $(this).hasClass('active')) {
                 $(this).removeClass('active');
                 $(this).find('[type="checkbox"]').removeAttr('checked');
             } else {
                 $(this).addClass('active');
-                $(this).find('[type="checkbox"]').attr('checked','true');
+                $(this).find('[type="checkbox"]').attr('checked', 'true');
             }
         });
 
@@ -206,41 +200,41 @@ export class WizardComponent implements OnInit, OnChanges, AfterViewInit{
 
     }
 
-    ngOnChanges(){
-        var input = $(this);
-        var target:EventTarget;
+    ngOnChanges(changes: SimpleChanges) {
+        const input = $(this);
+        const target: EventTarget = null;
         if (input.files && input.files[0]) {
-            var reader:any = new FileReader();
+            const reader: any = new FileReader();
 
             reader.onload = function (e) {
                 $('#wizardPicturePreview').attr('src', e.target.result).fadeIn('slow');
-            }
+            };
             reader.readAsDataURL(input.files[0]);
         }
     }
-    ngAfterViewInit(){
+    ngAfterViewInit() {
         $('.wizard-card').each(function(){
 
-            var $wizard = $(this);
-            var index = $wizard.bootstrapWizard('currentIndex');
+            const $wizard = $(this);
+            const index = $wizard.bootstrapWizard('currentIndex');
             // this.refreshAnimation($wizard, index);
 
-            var total_steps = $wizard.find('li').length;
-            var move_distance = $wizard.width() / total_steps;
-            var step_width = move_distance;
+            const total_steps = $wizard.find('li').length;
+            let move_distance = $wizard.width() / total_steps;
+            const step_width = move_distance;
             move_distance *= index;
 
-            var $current = index + 1;
+            const $current = index + 1;
 
-            if($current == 1){
+            if ($current === 1) {
                 move_distance -= 8;
-            } else if($current == total_steps){
+            } else if ($current === total_steps) {
                 move_distance += 8;
             }
 
             $wizard.find('.moving-tab').css('width', step_width);
             $('.moving-tab').css({
-                'transform':'translate3d(' + move_distance + 'px, 0, 0)',
+                'transform': 'translate3d(' + move_distance + 'px, 0, 0)',
                 'transition': 'all 0.5s cubic-bezier(0.29, 1.42, 0.79, 1)'
 
             });
