@@ -1,22 +1,23 @@
 import { Component, OnInit, ElementRef } from '@angular/core';
 import { Router, ActivatedRoute } from '@angular/router';
 
-declare var $:any;
+declare var $: any;
+
 @Component({
-    selector: 'my-app',
+    selector: 'app-my-app',
     templateUrl: './app.component.html'
 })
 
-export class AppComponent implements OnInit{
-    constructor(private elRef:ElementRef) {}
-    ngOnInit(){
-        let body = document.getElementsByTagName('body')[0];
-        var isWindows = navigator.platform.indexOf('Win') > -1 ? true : false;
-        if (isWindows){
+export class AppComponent implements OnInit {
+    constructor(private elRef: ElementRef) {}
+    ngOnInit() {
+        const body = document.getElementsByTagName('body')[0];
+        const isWindows = navigator.platform.indexOf('Win') > -1 ? true : false;
+        if (isWindows) {
            // if we are on windows OS we activate the perfectScrollbar function
-            body.classList.add("perfect-scrollbar-on");
+            body.classList.add('perfect-scrollbar-on');
         } else {
-            body.classList.add("perfect-scrollbar-off");
+            body.classList.add('perfect-scrollbar-off');
         }
         $.material.init();
     }
