@@ -1,27 +1,32 @@
 import { NgModule } from '@angular/core';
-import { BrowserModule } from '@angular/platform-browser';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations'; // this is needed!
 import { RouterModule } from '@angular/router';
 import { HttpModule } from '@angular/http';
 import { APP_BASE_HREF } from '@angular/common';
 import { FormsModule } from '@angular/forms';
+import { CommonModule } from '@angular/common';
 
-import { AppComponent }   from './app.component';
+import { AppComponent } from './app.component';
 
 import { SidebarModule } from './sidebar/sidebar.module';
 import { FooterModule } from './shared/footer/footer.module';
 import { NavbarModule} from './shared/navbar/navbar.module';
+import { FixedpluginModule } from './shared/fixedplugin/fixedplugin.module';
 import { AdminLayoutComponent } from './layouts/admin/admin-layout.component';
 import { AuthLayoutComponent } from './layouts/auth/auth-layout.component';
 import { AppRoutes } from './app.routing';
+
 @NgModule({
     imports:      [
-        BrowserModule,
+        CommonModule,
+        BrowserAnimationsModule,
         FormsModule,
         RouterModule.forRoot(AppRoutes),
         HttpModule,
         SidebarModule,
         NavbarModule,
-        FooterModule
+        FooterModule,
+        FixedpluginModule
     ],
     declarations: [
         AppComponent,
