@@ -16,31 +16,31 @@ export const AppRoutes: Routes = [
           {
         path: '',
         loadChildren: './dashboard/dashboard.module#DashboardModule'
-    },{
+    }, {
         path: 'components',
         loadChildren: './components/components.module#ComponentsModule'
-    },{
+    }, {
         path: 'forms',
         loadChildren: './forms/forms.module#Forms'
-    },{
+    }, {
         path: 'tables',
         loadChildren: './tables/tables.module#TablesModule'
-    },{
+    }, {
         path: 'maps',
         loadChildren: './maps/maps.module#MapsModule'
-    },{
+    }, {
         path: 'widgets',
         loadChildren: './widgets/widgets.module#WidgetsModule'
-    },{
+    }, {
         path: 'charts',
         loadChildren: './charts/charts.module#ChartsModule'
-    },{
+    }, {
         path: 'calendar',
         loadChildren: './calendar/calendar.module#CalendarModule'
-    },{
+    }, {
         path: '',
         loadChildren: './userpage/user.module#UserModule'
-    },{
+    }, {
         path: '',
         loadChildren: './timeline/timeline.module#TimelineModule'
     }
