@@ -1,147 +1,146 @@
 import { Component, OnInit } from '@angular/core';
-declare var swal: any;
-declare var $: any;
+declare const swal: any;
+declare const $: any;
 
 @Component({
-    moduleId: module.id,
-    selector: 'calendar-cmp',
+    selector: 'app-calendar-cmp',
     templateUrl: 'calendar.component.html'
 })
 
-export class CalendarComponent implements OnInit{
-    ngOnInit(){
-        var $calendar = $('#fullCalendar');
+export class CalendarComponent implements OnInit {
+    ngOnInit() {
+        const $calendar = $('#fullCalendar');
 
-        var today = new Date();
-        var y = today.getFullYear();
-        var m = today.getMonth();
-        var d = today.getDate();
+        const today = new Date();
+        const y = today.getFullYear();
+        const m = today.getMonth();
+        const d = today.getDate();
 
         $calendar.fullCalendar({
             viewRender: function(view, element) {
                 // We make sure that we activate the perfect scrollbar when the view isn't on Month
-                if (view.name != 'month'){
-                    var $fc_scroller = $('.fc-scroller');
+                if (view.name !== 'month') {
+                    const $fc_scroller = $('.fc-scroller');
                     $fc_scroller.perfectScrollbar();
                 }
             },
             header: {
-				left: 'title',
-				center: 'month,agendaWeek,agendaDay',
-				right: 'prev,next,today'
-			},
-			defaultDate: today,
-			selectable: true,
-			selectHelper: true,
+                left: 'title',
+                center: 'month, agendaWeek, agendaDay',
+                right: 'prev, next, today'
+            },
+            defaultDate: today,
+            selectable: true,
+            selectHelper: true,
             views: {
                 month: { // name of view
                     titleFormat: 'MMMM YYYY'
                     // other view-specific options here
                 },
                 week: {
-                    titleFormat: " MMMM D YYYY"
+                    titleFormat: ' MMMM D YYYY'
                 },
                 day: {
                     titleFormat: 'D MMM, YYYY'
                 }
             },
 
-			select: function(start, end) {
+            select: function(start, end) {
 
                 // on select we show the Sweet Alert modal with an input
-				swal({
-    				title: 'Create an Event',
-    				html: '<div class="form-group">' +
+                swal({
+                    title: 'Create an Event',
+                    html: '<div class="form-group">' +
                             '<input class="form-control" placeholder="Event Title" id="input-field">' +
                         '</div>',
-    				showCancelButton: true,
+                    showCancelButton: true,
                     confirmButtonClass: 'btn btn-success',
                     cancelButtonClass: 'btn btn-danger',
                     buttonsStyling: false
                 }).then(function(result) {
 
-                    var eventData;
-                    var event_title = $('#input-field').val();
+                    let eventData;
+                    const event_title = $('#input-field').val();
 
                     if (event_title) {
-    					eventData = {
-    						title: event_title,
-    						start: start,
-    						end: end
-    					};
-    					$calendar.fullCalendar('renderEvent', eventData, true); // stick? = true
-    				}
+                        eventData = {
+                            title: event_title,
+                            start: start,
+                            end: end
+                        };
+                        $calendar.fullCalendar('renderEvent', eventData, true); // stick? = true
+                    }
 
-    				$calendar.fullCalendar('unselect');
+                    $calendar.fullCalendar('unselect');
 
                 });
-			},
-			editable: true,
-			eventLimit: true, // allow "more" link when too many events
+            },
+            editable: true,
+            eventLimit: true, // allow "more" link when too many events
 
 
             // color classes: [ event-blue | event-azure | event-green | event-orange | event-red ]
             events: [
-				{
-					title: 'All Day Event',
-					start: new Date(y, m, 1),
+                {
+                    title: 'All Day Event',
+                    start: new Date(y, m, 1),
                     className: 'event-default'
-				},
-				{
-					id: 999,
-					title: 'Repeating Event',
-					start: new Date(y, m, d-4, 6, 0),
-					allDay: false,
-					className: 'event-rose'
-				},
-				{
-					id: 999,
-					title: 'Repeating Event',
-					start: new Date(y, m, d+3, 6, 0),
-					allDay: false,
-					className: 'event-rose'
-				},
-				{
-					title: 'Meeting',
-					start: new Date(y, m, d-1, 10, 30),
-					allDay: false,
-					className: 'event-green'
-				},
-				{
-					title: 'Lunch',
-					start: new Date(y, m, d+7, 12, 0),
-					end: new Date(y, m, d+7, 14, 0),
-					allDay: false,
-					className: 'event-red'
-				},
-				{
-					title: 'Md-pro Launch',
-					start: new Date(y, m, d-2, 12, 0),
-					allDay: true,
-					className: 'event-azure'
-				},
-				{
-					title: 'Birthday Party',
-					start: new Date(y, m, d+1, 19, 0),
-					end: new Date(y, m, d+1, 22, 30),
-					allDay: false,
+                },
+                {
+                    id: 999,
+                    title: 'Repeating Event',
+                    start: new Date(y, m, d - 4, 6, 0),
+                    allDay: false,
+                    className: 'event-rose'
+                },
+                {
+                    id: 999,
+                    title: 'Repeating Event',
+                    start: new Date(y, m, d + 3, 6, 0),
+                    allDay: false,
+                    className: 'event-rose'
+                },
+                {
+                    title: 'Meeting',
+                    start: new Date(y, m, d - 1, 10, 30),
+                    allDay: false,
+                    className: 'event-green'
+                },
+                {
+                    title: 'Lunch',
+                    start: new Date(y, m, d + 7, 12, 0),
+                    end: new Date(y, m, d + 7, 14, 0),
+                    allDay: false,
+                    className: 'event-red'
+                },
+                {
+                    title: 'Md-pro Launch',
+                    start: new Date(y, m, d - 2, 12, 0),
+                    allDay: true,
                     className: 'event-azure'
-				},
-				{
-					title: 'Click for Creative Tim',
-					start: new Date(y, m, 21),
-					end: new Date(y, m, 22),
-					url: 'https://www.creative-tim.com/',
-					className: 'event-orange'
-				},
-				{
-					title: 'Click for Google',
-					start: new Date(y, m, 21),
-					end: new Date(y, m, 22),
-					url: 'https://www.creative-tim.com/',
-					className: 'event-orange'
-				}
-			]
-		});
+                },
+                {
+                    title: 'Birthday Party',
+                    start: new Date(y, m, d + 1, 19, 0),
+                    end: new Date(y, m, d + 1, 22, 30),
+                    allDay: false,
+                    className: 'event-azure'
+                },
+                {
+                    title: 'Click for Creative Tim',
+                    start: new Date(y, m, 21),
+                    end: new Date(y, m, 22),
+                    url: 'https://www.creative-tim.com/',
+                    className: 'event-orange'
+                },
+                {
+                    title: 'Click for Google',
+                    start: new Date(y, m, 21),
+                    end: new Date(y, m, 22),
+                    url: 'https://www.creative-tim.com/',
+                    className: 'event-orange'
+                }
+            ]
+        });
     }
 }
