@@ -1,15 +1,8 @@
-import { Component, OnInit } from '@angular/core';
-
-declare var $:any;
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-buttons',
   templateUrl: './buttons.component.html'
 })
-export class ButtonsComponent implements OnInit {
 
-  // constructor(private navbarTitleService: NavbarTitleService, private notificationService: NotificationService) { }
-
-  public ngOnInit() {
-  }
-}
+export class ButtonsComponent {}
