@@ -15,37 +15,37 @@ export const ComponentsRoutes: Routes = [
       children: [ {
         path: 'buttons',
         component: ButtonsComponent
-    }]},{
+    }]}, {
     path: '',
     children: [ {
       path: 'grid',
       component: GridSystemComponent
     }]
-    },{
+    }, {
       path: '',
       children: [ {
         path: 'icons',
         component: IconsComponent
         }]
-    },{
+    }, {
         path: '',
         children: [ {
             path: 'notifications',
             component: NotificationsComponent
         }]
-    },{
+    }, {
         path: '',
         children: [ {
             path: 'panels',
             component: PanelsComponent
         }]
-    },{
+    }, {
         path: '',
         children: [ {
             path: 'sweet-alert',
             component: SweetAlertComponent
         }]
-    },{
+    }, {
         path: '',
         children: [ {
             path: 'typography',
