@@ -1,9 +1,8 @@
 import { Component } from '@angular/core';
 
 @Component({
-    moduleId: module.id,
-    selector: 'grid-cmp',
+    selector: 'app-grid-cmp',
     templateUrl: 'grid.component.html'
 })
 
-export class GridSystemComponent{}
+export class GridSystemComponent {}
