@@ -1,9 +1,8 @@
 import { Component } from '@angular/core';
 
 @Component({
-    moduleId: module.id,
-    selector: 'icons-cmp',
+    selector: 'app-icons-cmp',
     templateUrl: 'icons.component.html'
 })
 
-export class IconsComponent{}
+export class IconsComponent {}
