@@ -1,23 +1,21 @@
-import { Component, OnInit } from '@angular/core';
+import { Component } from '@angular/core';
 
-declare var $:any;
+declare const $: any;
 
 @Component({
-    moduleId: module.id,
-    selector: 'notifications-cmp',
+    selector: 'app-notifications-cmp',
     templateUrl: 'notifications.component.html'
 })
-export class NotificationsComponent implements OnInit{
-    showNotification(from, align){
-        var type = ['','info','success','warning','danger','rose','primary'];
+export class NotificationsComponent {
+    showNotification(from, align) {
+        const type = ['', 'info', 'success', 'warning', 'danger', 'rose', 'primary'];
 
-        var color = Math.floor((Math.random() * 6) + 1);
+        const color = Math.floor((Math.random() * 6) + 1);
 
-    	$.notify({
-        	icon: "notifications",
-        	message: "Welcome to <b>Material Dashboard</b> - a beautiful dashboard for every web developer."
-
-        },{
+        $.notify({
+            icon: 'notifications',
+            message: 'Welcome to <b>Material Dashboard</b> - a beautiful dashboard for every web developer.'
+        }, {
             type: type[color],
             timer: 3000,
             placement: {
@@ -25,9 +23,5 @@ export class NotificationsComponent implements OnInit{
                 align: align
             }
         });
-	}
-    ngOnInit(){
-        // We put modals out of wrapper to working properly
-        // $('.modal').appendTo("body");
     }
 }
