@@ -1,45 +1,43 @@
-import { Component, OnInit } from '@angular/core';
+import { Component } from '@angular/core';
 
-// import swal from 'sweetalert2'
-declare var $:any;
-declare var swal:any;
+declare var $: any;
+declare var swal: any;
+
 @Component({
-    moduleId: module.id,
-    selector: 'sweetalert-cmp',
+    selector: 'app-sweetalert-cmp',
     templateUrl: 'sweetalert.component.html'
 })
 
-export class SweetAlertComponent implements OnInit{
-    showSwal(type){
-        if(type == 'basic'){
-        	swal({
-                title: "Here's a message!",
+export class SweetAlertComponent {
+    showSwal(type) {
+        if (type === 'basic') {
+            swal({
+                title: 'Here is a message!',
                 buttonsStyling: false,
-                confirmButtonClass: "btn btn-success"
+                confirmButtonClass: 'btn btn-success'
+            });
+        } else if (type === 'title-and-text') {
+            swal({
+                title: 'Here is a message!',
+                text: 'It is pretty, is not it?',
+                buttonsStyling: false,
+                confirmButtonClass: 'btn btn-info'
             });
 
-    	}else if(type == 'title-and-text'){
-        	swal({
-                title: "Here's a message!",
-                text: "It's pretty, isn't it?",
+        } else if (type === 'success-message') {
+            swal({
+                type: 'success',
+                title: 'Good job!',
+                text: 'You clicked the button!',
                 buttonsStyling: false,
-                confirmButtonClass: "btn btn-info"
-            });
-
-    	}else if(type == 'success-message'){
-        	swal({
-                type: "success",
-                title: "Good job!",
-                text: "You clicked the button!",
-                buttonsStyling: false,
-                confirmButtonClass: "btn btn-success"
+                confirmButtonClass: 'btn btn-success'
 
             });
 
-    	}else if(type == 'warning-message-and-confirmation'){
+        } else if (type === 'warning-message-and-confirmation') {
             swal({
                     title: 'Are you sure?',
-                    text: "You won't be able to revert this!",
+                    text: 'You will not be able to revert this!',
                     type: 'warning',
                     showCancelButton: true,
                     confirmButtonClass: 'btn btn-success',
@@ -51,11 +49,11 @@ export class SweetAlertComponent implements OnInit{
                     title: 'Deleted!',
                     text: 'Your file has been deleted.',
                     type: 'success',
-                    confirmButtonClass: "btn btn-success",
+                    confirmButtonClass: 'btn btn-success',
                     buttonsStyling: false
-                    })
                 });
-    	}else if(type == 'warning-message-and-cancel'){
+                });
+            } else if (type === 'warning-message-and-cancel') {
             swal({
                     title: 'Are you sure?',
                     text: 'You will not be able to recover this imaginary file!',
@@ -63,17 +61,17 @@ export class SweetAlertComponent implements OnInit{
                     showCancelButton: true,
                     confirmButtonText: 'Yes, delete it!',
                     cancelButtonText: 'No, keep it',
-                    confirmButtonClass: "btn btn-success",
-                    cancelButtonClass: "btn btn-danger",
+                    confirmButtonClass: 'btn btn-success',
+                    cancelButtonClass: 'btn btn-danger',
                     buttonsStyling: false
                 }).then(function() {
                   swal({
                     title: 'Deleted!',
                     text: 'Your imaginary file has been deleted.',
                     type: 'success',
-                    confirmButtonClass: "btn btn-success",
+                    confirmButtonClass: 'btn btn-success',
                     buttonsStyling: false
-                    })
+                });
                 }, function(dismiss) {
                   // dismiss can be 'overlay', 'cancel', 'close', 'esc', 'timer'
                   if (dismiss === 'cancel') {
@@ -81,30 +79,29 @@ export class SweetAlertComponent implements OnInit{
                       title: 'Cancelled',
                       text: 'Your imaginary file is safe :)',
                       type: 'error',
-                      confirmButtonClass: "btn btn-info",
+                      confirmButtonClass: 'btn btn-info',
                       buttonsStyling: false
-                    })
+                  });
                   }
-                })
-
-    	}else if(type == 'custom-html'){
-        	swal({
+              });
+          } else if (type === 'custom-html') {
+            swal({
                 title: 'HTML example',
                 buttonsStyling: false,
-                confirmButtonClass: "btn btn-success",
+                confirmButtonClass: 'btn btn-success',
                 html:
                         'You can use <b>bold text</b>, ' +
                         '<a href="http://github.com">links</a> ' +
                         'and other HTML tags'
                 });
 
-    	}else if(type == 'auto-close'){
-        	swal({ title: "Auto close alert!",
-            	   text: "I will close in 2 seconds.",
-            	   timer: 2000,
-            	   showConfirmButton: false
+            } else if (type === 'auto-close') {
+            swal({ title: 'Auto close alert!',
+            text: 'I will close in 2 seconds.',
+                   timer: 2000,
+                   showConfirmButton: false
                 });
-    	} else if(type == 'input-field'){
+            } else if (type === 'input-field') {
             swal({
                     title: 'Input something',
                     html: '<div class="form-group">' +
@@ -123,9 +120,8 @@ export class SweetAlertComponent implements OnInit{
                         confirmButtonClass: 'btn btn-success',
                         buttonsStyling: false
 
-                    })
-                }).catch(swal.noop)
+                    });
+                }).catch(swal.noop);
         }
     }
-    public ngOnInit(){}
 }
