@@ -1,9 +1,8 @@
 import { Component } from '@angular/core';
 
 @Component({
-    moduleId: module.id,
-    selector: 'typography-cmp',
+    selector: 'app-typography-cmp',
     templateUrl: 'typography.component.html'
 })
 
-export class TypographyComponent{}
+export class TypographyComponent {}
