@@ -1,25 +1,26 @@
 import { Component, OnInit, AfterViewInit } from '@angular/core';
 import { TableData } from '../md/md-table/md-table.component';
+import { LegendItem, ChartType } from '../md/md-chart/md-chart.component';
 
 import * as Chartist from 'chartist';
 
-declare var $:any;
+declare const $: any;
 
 @Component({
   selector: 'app-dashboard',
   templateUrl: './dashboard.component.html'
 })
-export class DashboardComponent implements OnInit, AfterViewInit{
+export class DashboardComponent implements OnInit, AfterViewInit {
   // constructor(private navbarTitleService: NavbarTitleService, private notificationService: NotificationService) { }
   public tableData: TableData;
-  startAnimationForLineChart(chart){
-      var seq, delays, durations;
+  startAnimationForLineChart(chart: any) {
+      let seq: any, delays: any, durations: any;
       seq = 0;
       delays = 80;
       durations = 500;
-      chart.on('draw', function(data) {
+      chart.on('draw', function(data: any) {
 
-        if(data.type === 'line' || data.type === 'area') {
+        if (data.type === 'line' || data.type === 'area') {
           data.element.animate({
             d: {
               begin: 600,
@@ -29,7 +30,7 @@ export class DashboardComponent implements OnInit, AfterViewInit{
               easing: Chartist.Svg.Easing.easeOutQuint
             }
           });
-        } else if(data.type === 'point') {
+        } else if (data.type === 'point') {
               seq++;
               data.element.animate({
                 opacity: {
@@ -45,13 +46,13 @@ export class DashboardComponent implements OnInit, AfterViewInit{
 
       seq = 0;
   }
-  startAnimationForBarChart(chart){
-      var seq2, delays2, durations2;
+  startAnimationForBarChart(chart: any) {
+      let seq2: any, delays2: any, durations2: any;
       seq2 = 0;
       delays2 = 80;
       durations2 = 500;
-      chart.on('draw', function(data) {
-        if(data.type === 'bar'){
+      chart.on('draw', function(data: any) {
+        if (data.type === 'bar') {
             seq2++;
             data.element.animate({
               opacity: {
@@ -82,57 +83,59 @@ export class DashboardComponent implements OnInit, AfterViewInit{
        };
       /* ----------==========     Daily Sales Chart initialization    ==========---------- */
 
-      var dataDailySalesChart = {
+      const dataDailySalesChart = {
           labels: ['M', 'T', 'W', 'T', 'F', 'S', 'S'],
           series: [
               [12, 17, 7, 17, 23, 18, 38]
           ]
       };
 
-     var optionsDailySalesChart = {
+     const optionsDailySalesChart = {
           lineSmooth: Chartist.Interpolation.cardinal({
               tension: 0
           }),
           low: 0,
           high: 50, // creative tim: we recommend you to set the high sa the biggest value + something for a better look
           chartPadding: { top: 0, right: 0, bottom: 0, left: 0},
-      }
+      };
 
-      var dailySalesChart = new Chartist.Line('#dailySalesChart', dataDailySalesChart, optionsDailySalesChart);
+      const dailySalesChart = new Chartist.Line('#dailySalesChart', dataDailySalesChart, optionsDailySalesChart);
 
       this.startAnimationForLineChart(dailySalesChart);
       /* ----------==========     Completed Tasks Chart initialization    ==========---------- */
 
-      var dataCompletedTasksChart = {
+      const dataCompletedTasksChart = {
           labels: ['12p', '3p', '6p', '9p', '12p', '3a', '6a', '9a'],
           series: [
               [230, 750, 450, 300, 280, 240, 200, 190]
           ]
       };
 
-      var optionsCompletedTasksChart = {
+      const optionsCompletedTasksChart = {
           lineSmooth: Chartist.Interpolation.cardinal({
               tension: 0
           }),
           low: 0,
-          high: 1000, // creative tim: we recommend you to set the high sa the biggest value + something for a better look
+          high: 1000, // creative tim: we recommend you to set the high sa the biggest value + something for a better
+          // look
           chartPadding: { top: 0, right: 0, bottom: 0, left: 0}
-      }
+      };
 
-     var completedTasksChart = new Chartist.Line('#completedTasksChart', dataCompletedTasksChart, optionsCompletedTasksChart);
+     const completedTasksChart = new Chartist.Line('#completedTasksChart', dataCompletedTasksChart,
+      optionsCompletedTasksChart);
 
      this.startAnimationForLineChart(completedTasksChart);
 
       /* ----------==========     Emails Subscription Chart initialization    ==========---------- */
 
-      var dataWebsiteViewsChart = {
+      const dataWebsiteViewsChart = {
         labels: ['J', 'F', 'M', 'A', 'M', 'J', 'J', 'A', 'S', 'O', 'N', 'D'],
         series: [
           [542, 443, 320, 780, 553, 453, 326, 434, 568, 610, 756, 895]
 
         ]
       };
-      var optionsWebsiteViewsChart = {
+      const optionsWebsiteViewsChart = {
           axisX: {
               showGrid: false
           },
@@ -140,7 +143,7 @@ export class DashboardComponent implements OnInit, AfterViewInit{
           high: 1000,
           chartPadding: { top: 0, right: 5, bottom: 0, left: 0}
       };
-      var responsiveOptions:any = [
+      const responsiveOptions: any = [
         ['screen and (max-width: 640px)', {
           seriesBarDistance: 5,
           axisX: {
@@ -150,70 +153,69 @@ export class DashboardComponent implements OnInit, AfterViewInit{
           }
         }]
       ];
-      var websiteViewsChart = new Chartist.Bar('#websiteViewsChart', dataWebsiteViewsChart, optionsWebsiteViewsChart, responsiveOptions);
+      const websiteViewsChart = new Chartist.Bar('#websiteViewsChart', dataWebsiteViewsChart, optionsWebsiteViewsChart, responsiveOptions);
 
       this.startAnimationForBarChart(websiteViewsChart);
 
-      var mapData = {
-           "AU": 760,
-           "BR": 550,
-           "CA": 120,
-           "DE": 1300,
-           "FR": 540,
-           "GB": 690,
-           "GE": 200,
-           "IN": 200,
-           "RO": 600,
-           "RU": 300,
-           "US": 2920,
+      const mapData = {
+           'AU': 760,
+           'BR': 550,
+           'CA': 120,
+           'DE': 1300,
+           'FR': 540,
+           'GB': 690,
+           'GE': 200,
+           'IN': 200,
+           'RO': 600,
+           'RU': 300,
+           'US': 2920,
        };
           $('#worldMap').vectorMap({
               map: 'world_mill_en',
-              backgroundColor: "transparent",
+              backgroundColor: 'transparent',
               zoomOnScroll: false,
               regionStyle: {
                   initial: {
                       fill: '#e4e4e4',
-                      "fill-opacity": 0.9,
+                      'fill-opacity': 0.9,
                       stroke: 'none',
-                      "stroke-width": 0,
-                      "stroke-opacity": 0
+                      'stroke-width': 0,
+                      'stroke-opacity': 0
                   }
               },
 
               series: {
                   regions: [{
                       values: mapData,
-                      scale: ["#AAAAAA","#444444"],
+                      scale: ['#AAAAAA', '#444444'],
                       normalizeFunction: 'polynomial'
                   }]
               },
           });
    }
-   ngAfterViewInit(){
-       var breakCards = true;
-       if(breakCards == true){
+   ngAfterViewInit() {
+       const breakCards = true;
+       if (breakCards === true) {
            // We break the cards headers if there is too much stress on them :-)
            $('[data-header-animation="true"]').each(function(){
-               var $fix_button = $(this);
-               var $card = $(this).parent('.card');
+               const $fix_button = $(this);
+               const $card = $(this).parent('.card');
                $card.find('.fix-broken-card').click(function(){
-                   console.log(this);
-                   var $header = $(this).parent().parent().siblings('.card-header, .card-image');
+                   const $header = $(this).parent().parent().siblings('.card-header, .card-image');
                    $header.removeClass('hinge').addClass('fadeInDown');
 
-                   $card.attr('data-count',0);
+                   $card.attr('data-count', 0);
 
                    setTimeout(function(){
                        $header.removeClass('fadeInDown animate');
-                   },480);
+                   }, 480);
                });
 
                $card.mouseenter(function(){
-                   var $this = $(this);
-                   var hover_count = parseInt($this.attr('data-count'), 10) + 1 || 0;
-                   $this.attr("data-count", hover_count);
-                   if (hover_count >= 20){
+                   const $this = $(this);
+                   const hover_count = parseInt($this.attr('data-count'), 10) + 1 || 0;
+                   $this.attr('data-count', hover_count);
+                   if (hover_count >= 20) {
                        $(this).children('.card-header, .card-image').addClass('hinge animated');
                    }
                });
