@@ -11,19 +11,19 @@ export const FormsRoutes: Routes = [
       children: [ {
         path: 'regular',
         component: RegularFormsComponent
-    }]},{
+    }]}, {
     path: '',
     children: [ {
       path: 'extended',
       component: ExtendedFormsComponent
     }]
-    },{
+    }, {
     path: '',
     children: [ {
       path: 'validation',
       component: ValidationFormsComponent
     }]
-    },{
+    }, {
         path: '',
         children: [ {
             path: 'wizard',
