@@ -15,27 +15,11 @@ declare interface User {
 }
 
 @Component({
-    moduleId: module.id,
-    selector: 'validationforms-cmp',
+    selector: 'app-validationforms-cmp',
     templateUrl: 'validationforms.component.html'
 })
 
-export class ValidationFormsComponent{
-    // rangeValidation : FormGroup;
-    //
-    // // We are passing an instance of the FormBuilder to our constructor
-    // constructor(fb: FormBuilder){
-    //   // Here we are using the FormBuilder to build out our form.
-    //   this.rangeValidation = fb.group({
-    //     // To add a validator, we must first convert the string value into an array. The first item in the array is the default value if any, then the next item in the array is the validator. Here we are adding a required validator meaning that the firstName attribute must have a value in it.
-    //     // To add a validator, we must first convert the string value into an array. The first item in the array is the default value if any, then the next item in the array is the validator. Here we are adding a required validator meaning that the firstName attribute must have a value in it.
-    //   'firstName' : [null, Validators.required],
-    //   // We can use more than one validator per field. If we want to use more than one validator we have to wrap our array of validators with a Validators.compose function. Here we are using a required, minimum length and maximum length validator.
-    //   'lastName': [null, Validators.compose([Validators.required, Validators.minLength(5), Validators.maxLength(10)])],
-    //   'gender' : [null, Validators.required],
-    //
-    //   })
-    // }
+export class ValidationFormsComponent implements OnInit {
     public user: User;
     public typeValidation: User;
 
@@ -44,21 +28,21 @@ export class ValidationFormsComponent{
       email: '',
       password: '',
       confirmPassword: ''
-    }
+  };
     this.typeValidation = {
         text: '',
         email: '',
         idSource: '',
         idDestination: '',
         url: ''
-    }
+    };
   }
 
   save(model: User, isValid: boolean) {
     // call API to save customer
     console.log(model, isValid);
   }
-    onSubmit(value: any):void{
+    onSubmit(value: any): void {
         console.log(value);
     }
 }
