@@ -1,5 +1,5 @@
 import { Routes } from '@angular/router';
-//
+
 import { FullScreenMapsComponent } from './fullscreenmap/fullscreenmap.component';
 import { GoogleMapsComponent } from './googlemaps/googlemaps.component';
 import { VectorMapsComponent } from './vectormaps/vectormaps.component';
@@ -11,13 +11,13 @@ export const MapsRoutes: Routes = [
         path: 'fullscreen',
         component: FullScreenMapsComponent
     }]
-    },{
+    }, {
     path: '',
     children: [ {
       path: 'google',
       component: GoogleMapsComponent
     }]
-    },{
+    }, {
       path: '',
       children: [ {
         path: 'vector',
