@@ -4,6 +4,7 @@ import { CommonModule } from '@angular/common';
 import { RouterModule } from '@angular/router';
 
 import { MdTableComponent } from './md-table/md-table.component';
+import { MdChartComponent } from './md-chart/md-chart.component';
 
 export interface DropdownLink {
   title: string;
@@ -32,10 +33,12 @@ export interface NavItem {
     RouterModule
   ],
   declarations: [
-    MdTableComponent
+    MdTableComponent,
+    MdChartComponent
   ],
   exports: [
     MdTableComponent,
+    MdChartComponent
   ]
 })
 export class MdModule { }
