@@ -1,31 +1,56 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, ElementRef } from '@angular/core';
 
-declare var $:any;
+declare var $: any;
 
 @Component({
-    moduleId:module.id,
-    selector: 'login-cmp',
+    selector: 'app-login-cmp',
     templateUrl: './login.component.html'
 })
 
-export class LoginComponent implements OnInit{
-    test : Date = new Date();
+export class LoginComponent implements OnInit {
+    test: Date = new Date();
+    private toggleButton: any;
+    private sidebarVisible: boolean;
+    private nativeElement: Node;
 
-    checkFullPageBackgroundImage(){
-        var $page = $('.full-page');
-        var image_src = $page.data('image');
+    constructor(private element: ElementRef) {
+        this.nativeElement = element.nativeElement;
+        this.sidebarVisible = false;
+    }
+    checkFullPageBackgroundImage() {
+        const $page = $('.full-page');
+        const image_src = $page.data('image');
 
-        if(image_src !== undefined){
-            var image_container = '<div class="full-page-background" style="background-image: url(' + image_src + ') "/>'
+        if (image_src !== undefined) {
+            const image_container = '<div class="full-page-background" style="background-image: url(' + image_src + ') "/>';
             $page.append(image_container);
         }
     };
-    ngOnInit(){
+    ngOnInit() {
         this.checkFullPageBackgroundImage();
 
-        setTimeout(function(){
+        var navbar : HTMLElement = this.element.nativeElement;
+        this.toggleButton = navbar.getElementsByClassName('navbar-toggle')[0];
+
+        setTimeout(function() {
             // after 1000 ms we add the class animated to the login/register card
             $('.card').removeClass('card-hidden');
-        }, 700)
+        }, 700);
+    }
+    sidebarToggle() {
+        var toggleButton = this.toggleButton;
+        var body = document.getElementsByTagName('body')[0];
+        var sidebar = document.getElementsByClassName('navbar-collapse')[0];
+        if (this.sidebarVisible == false) {
+            setTimeout(function() {
+                toggleButton.classList.add('toggled');
+            }, 500);
+            body.classList.add('nav-open');
+            this.sidebarVisible = true;
+        } else {
+            this.toggleButton.classList.remove('toggled');
+            this.sidebarVisible = false;
+            body.classList.remove('nav-open');
+        }
     }
 }
