@@ -12,13 +12,13 @@ export const PagesRoutes: Routes = [
         children: [ {
             path: 'login',
             component: LoginComponent
-        },{
+        }, {
             path: 'lock',
             component: LockComponent
-        },{
+        }, {
             path: 'register',
             component: RegisterComponent
-        },{
+        }, {
             path: 'pricing',
             component: PricingComponent
         }]
