@@ -1,26 +1,25 @@
 import { Component, OnInit } from '@angular/core';
 
-declare var $:any;
+declare var $: any;
 
 @Component({
-    moduleId:module.id,
-    selector: 'register-cmp',
+    selector: 'app-register-cmp',
     templateUrl: './register.component.html'
 })
 
-export class RegisterComponent implements OnInit{
-    test : Date = new Date();
+export class RegisterComponent implements OnInit {
+    test: Date = new Date();
 
-    checkFullPageBackgroundImage(){
-        var $page = $('.full-page');
-        var image_src = $page.data('image');
+    checkFullPageBackgroundImage() {
+        const $page = $('.full-page');
+        const image_src = $page.data('image');
 
-        if(image_src !== undefined){
-            var image_container = '<div class="full-page-background" style="background-image: url(' + image_src + ') "/>'
+        if (image_src !== undefined) {
+            const image_container = '<div class="full-page-background" style="background-image: url(' + image_src + ') "/>';
             $page.append(image_container);
         }
     };
-    ngOnInit(){
+    ngOnInit() {
         this.checkFullPageBackgroundImage();
     }
 }
