@@ -1,11 +1,10 @@
 import { Component } from '@angular/core';
 
 @Component({
-    moduleId: module.id,
-    selector: 'footer-cmp',
+    selector: 'app-footer-cmp',
     templateUrl: 'footer.component.html'
 })
 
-export class FooterComponent{
-    test : Date = new Date();
+export class FooterComponent {
+    test: Date = new Date();
 }
