@@ -3,42 +3,38 @@ import { RouteInfo } from './sidebar.metadata';
 export const ROUTES: RouteInfo[] = [
     { path: '/dashboard', title: 'Dashboard', icon: 'material-icons' },
 
-    { path: '/pages/timeline', title: 'Timeline Page', icon:'material-icons' },
-    { path: '/pages/user', title: 'User Page', icon:'material-icons' },
+    { path: '/pages/timeline', title: 'Timeline Page', icon: 'material-icons' },
+    { path: '/pages/user', title: 'User Page', icon: 'material-icons' },
 
-    { path: '/components/buttons', title: 'Buttons', icon:'pe-7s-plugin' },
-    { path: '/components/grid', title: 'Grid System', icon:'pe-7s-plugin' },
-    { path: '/components/panels', title: 'Panels', icon:'pe-7s-plugin' },
-    { path: '/components/sweet-alert', title: 'Sweet Alert', icon:'pe-7s-plugin' },
-    { path: '/components/notifications', title: 'Notifications', icon:'pe-7s-plugin' },
-    { path: '/components/icons', title: 'Icons', icon:'pe-7s-plugin' },
-    { path: '/components/typography', title: 'Typography', icon:'pe-7s-plugin' },
+    { path: '/components/buttons', title: 'Buttons', icon: 'pe-7s-plugin' },
+    { path: '/components/grid', title: 'Grid System', icon: 'pe-7s-plugin' },
+    { path: '/components/panels', title: 'Panels', icon: 'pe-7s-plugin' },
+    { path: '/components/sweet-alert', title: 'Sweet Alert', icon: 'pe-7s-plugin' },
+    { path: '/components/notifications', title: 'Notifications', icon: 'pe-7s-plugin' },
+    { path: '/components/icons', title: 'Icons', icon: 'pe-7s-plugin' },
+    { path: '/components/typography', title: 'Typography', icon: 'pe-7s-plugin' },
 
-    { path: '/forms/regular', title: 'Regular Forms', icon:'pe-7s-note2' },
-    { path: '/forms/extended', title: 'Extended Forms', icon:'pe-7s-note2' },
-    { path: '/forms/validation', title: 'Validation Forms', icon:'pe-7s-note2' },
-    { path: '/forms/wizard', title: 'Wizard', icon:'pe-7s-note2' },
+    { path: '/forms/regular', title: 'Regular Forms', icon: 'pe-7s-note2' },
+    { path: '/forms/extended', title: 'Extended Forms', icon: 'pe-7s-note2' },
+    { path: '/forms/validation', title: 'Validation Forms', icon: 'pe-7s-note2' },
+    { path: '/forms/wizard', title: 'Wizard', icon: 'pe-7s-note2' },
 
-    { path: '/tables/regular', title: 'Regular Tables', icon:'pe-7s-news-paper' },
-    { path: '/tables/extended', title: 'Extended Tables', icon:'pe-7s-news-paper' },
-    { path: '/tables/datatables.net', title: 'DataTables.net', icon:'pe-7s-news-paper' },
+    { path: '/tables/regular', title: 'Regular Tables', icon: 'pe-7s-news-paper' },
+    { path: '/tables/extended', title: 'Extended Tables', icon: 'pe-7s-news-paper' },
+    { path: '/tables/datatables.net', title: 'DataTables.net', icon: 'pe-7s-news-paper' },
 
-    { path: '/maps/google', title: 'Google Maps', icon:'pe-7s-map-marker' },
-    { path: '/maps/fullscreen', title: 'Full Screen Map', icon:'pe-7s-map-marker' },
-    { path: '/maps/vector', title: 'Vector Map', icon:'pe-7s-map-marker' },
+    { path: '/maps/google', title: 'Google Maps', icon: 'pe-7s-map-marker' },
+    { path: '/maps/fullscreen', title: 'Full Screen Map', icon: 'pe-7s-map-marker' },
+    { path: '/maps/vector', title: 'Vector Map', icon: 'pe-7s-map-marker' },
 
-    { path: '/widgets', title: 'Widgets', icon:'material-icons' },
+    { path: '/widgets', title: 'Widgets', icon: 'material-icons' },
 
-    { path: '/charts', title: 'Charts', icon:'material-icons' },
+    { path: '/charts', title: 'Charts', icon: 'material-icons' },
 
-    { path: '/calendar', title: 'Calendar', icon:'material-icons' },
+    { path: '/calendar', title: 'Calendar', icon: 'material-icons' },
 
-    { path: '/pages/pricing', title: 'Pricing', icon:'material-icons' },
-    { path: '/pages/login', title: 'Login Page', icon:'material-icons' },
-    { path: '/pages/register', title: 'Register Page', icon:'material-icons' },
-    { path: '/pages/lock', title: 'Lock Screen Page', icon:'material-icons' },
-
-
-
-
+    { path: '/pages/pricing', title: 'Pricing', icon: 'material-icons' },
+    { path: '/pages/login', title: 'Login Page', icon: 'material-icons' },
+    { path: '/pages/register', title: 'Register Page', icon: 'material-icons' },
+    { path: '/pages/lock', title: 'Lock Screen Page', icon: 'material-icons' }
 ];
