@@ -1,9 +1,8 @@
 import { Component } from '@angular/core';
 
 @Component({
-    moduleId: module.id,
-    selector: 'timeline-cmp',
+    selector: 'app-timeline-cmp',
     templateUrl: 'timeline.component.html'
 })
 
-export class TimelineComponent{}
+export class TimelineComponent {}
