@@ -1,50 +1,8 @@
-import { Component,state,style,animate,transition, trigger, keyframes } from '@angular/core';
+import { Component } from '@angular/core';
 
 @Component({
-    moduleId: module.id,
-    selector: 'user-cmp',
-    templateUrl: 'user.component.html',
-    animations: [
-        trigger('carduserprofile', [
-            state('*', style({
-                '-ms-transform': 'translate3D(0px, 0px, 0px)',
-                '-webkit-transform': 'translate3D(0px, 0px, 0px)',
-                '-moz-transform': 'translate3D(0px, 0px, 0px)',
-                '-o-transform':'translate3D(0px, 0px, 0px)',
-                transform:'translate3D(0px, 0px, 0px)',
-                opacity: 1
-            })),
-            transition('void => *', [
-                style({opacity: 0,
-                    '-ms-transform': 'translate3D(0px, 150px, 0px)',
-                    '-webkit-transform': 'translate3D(0px, 150px, 0px)',
-                    '-moz-transform': 'translate3D(0px, 150px, 0px)',
-                    '-o-transform':'translate3D(0px, 150px, 0px)',
-                    transform:'translate3D(0px, 150px, 0px)',
-                }),
-                animate('0.3s 0s ease-out'),
-            ])
-        ]),
-        trigger('cardprofile', [
-            state('*', style({
-                '-ms-transform': 'translate3D(0px, 0px, 0px)',
-                '-webkit-transform': 'translate3D(0px, 0px, 0px)',
-                '-moz-transform': 'translate3D(0px, 0px, 0px)',
-                '-o-transform':'translate3D(0px, 0px, 0px)',
-                transform:'translate3D(0px, 0px, 0px)',
-                opacity: 1})),
-                transition('void => *', [
-                    style({opacity: 0,
-                        '-ms-transform': 'translate3D(0px, 150px, 0px)',
-                        '-webkit-transform': 'translate3D(0px, 150px, 0px)',
-                        '-moz-transform': 'translate3D(0px, 150px, 0px)',
-                        '-o-transform':'translate3D(0px, 150px, 0px)',
-                        transform:'translate3D(0px, 150px, 0px)',
-                    }),
-                    animate('0.3s 0.25s ease-out')
-                ])
-            ])
-        ]
-    })
+    selector: 'app-user-cmp',
+    templateUrl: 'user.component.html'
+})
 
-    export class UserComponent{ }
+export class UserComponent {}
